pyproject.toml
[project]
name = "sensor-simulant"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergey Malinovskiy",email = "malinovskiy@firecode-team.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "sensor_simulant", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
